* Create multi_multi branch, do get multi-platform multi-player working  * moved Projectile and Splash ParticleSystems from Shared::Graphics into Glest::Game  * Removed Shared::Graphics::Entity  * Removed Shared::Graphics::ParticleObsrerver  * renamed GameNetworkInterface => GameInterface  * All 'updates' that could cause sync issues in multiplayer are now routed though GameInterface  * Moved some constants from Unit into GameConstants  * Moved some constants from NetworkInterface into GameConstants  * Shared:Util::Conversion is now a namespace, not a class  * WorldUpdateFps is now in GameConstants, the config option is only used in debug edition  * Added SkillCycleTable, server pre-calculates skill cycle lengths and sends to clients before game starts  * Added KeyFrame, is used to send commands from the server instead of a NetworkMessageCommandList, currently also contains shitloads of checksums  * introduced Shared::Math::fixed, to replace floats for anything that can effect game state  * Mass changes of floating point to fixed point calculation  * removed all instances of 'using namespace std;' added using declarations for most common stuff to pch.h  * renamed UnitStatsBase => UnitStats, & EnhancementTypeBase => EnhancementType  * moved some iterators from map.h into pos_iterator.h  * moved earthquake stuff out of map.h/.cpp into its own files  * probably some other stuff ?

+++ source/game/gui/gui.cpp
@@ -1213,13 +1213,13 @@ void Gui::computeBuildPositions(const Vec2i &end) {
 	if(abs(offset.x) > abs(offset.y)) {
 		count = abs(offset.x / size) + 1;
 		offsetAdjusted.x = (offset.x / size) * size;
-		float mulit = (float)offset.x / (float)offsetAdjusted.x;
-		offsetAdjusted.y = (float)offset.y * mulit;
+		float mulit = float(offset.x) / float(offsetAdjusted.x);
+		offsetAdjusted.y = int(float(offset.y) * mulit);
 	} else {
 		count = abs(offset.y / size) + 1;
 		offsetAdjusted.y = (offset.y / size) * size;
-		float mulit = (float)offset.y / (float)offsetAdjusted.y;
-		offsetAdjusted.x = (float)offset.x * mulit;
+		float mulit = float(offset.y) / float(offsetAdjusted.y);
+		offsetAdjusted.x = int(float(offset.x * mulit));
 	}
 
 	buildPositions.resize(count);
